implement passing contextPath

+++ src/prisma-generator.ts
@@ -1,13 +1,19 @@
-import { parseEnvValue, getDMMF } from '@prisma/internals';
 import { EnvValue, GeneratorOptions } from '@prisma/generator-helper';
-import path from 'path';
+import { getDMMF, parseEnvValue } from '@prisma/internals';
 import { promises as fs } from 'fs';
+import path from 'path';
+import { configSchema } from './config';
+import { constructShield } from './helpers';
+import { RootType } from './types';
 import removeDir from './utils/removeDir';
 import { writeFileSafely } from './utils/writeFileSafely';
-import { constructShield } from './helpers';
 
 export async function generate(options: GeneratorOptions) {
   const outputDir = parseEnvValue(options.generator.output as EnvValue);
+  const results = configSchema.safeParse(options.generator.config);
+  if (!results.success) throw new Error('Invalid options passed');
+  const config = results.data;
+
   await fs.mkdir(outputDir, { recursive: true });
   await removeDir(outputDir, true);
 
@@ -20,9 +26,9 @@ export async function generate(options: GeneratorOptions) {
     previewFeatures: prismaClientProvider?.previewFeatures,
   });
 
-  const queries: Array<string> = [];
-  const mutations: Array<string> = [];
-  const subscriptions: Array<string> = [];
+  const queries: RootType = [];
+  const mutations: RootType = [];
+  const subscriptions: RootType = [];
 
   prismaClientDmmf.mappings.modelOperations.forEach((modelOperation) => {
     const { model, plural, ...operations } = modelOperation;
@@ -36,7 +42,7 @@ export async function generate(options: GeneratorOptions) {
           'groupBy',
         ].includes(opType)
       ) {
-        queries.push(opNameWithModel);
+        queries.push(opNameWithModel as string);
       }
 
       if (
@@ -49,7 +55,7 @@ export async function generate(options: GeneratorOptions) {
           'upsertOne',
         ].includes(opType)
       ) {
-        mutations.push(opNameWithModel);
+        mutations.push(opNameWithModel as string);
       }
     }
   });
@@ -57,6 +63,6 @@ export async function generate(options: GeneratorOptions) {
   queries.sort();
   mutations.sort();
   subscriptions.sort();
-  const shieldText = constructShield({ queries, mutations, subscriptions });
+  const shieldText = constructShield({ queries, mutations, subscriptions }, config, options);
   await writeFileSafely(path.join(outputDir, 'shield.ts'), shieldText);
 }
